pyproject.toml
[project]
name = "safe-env"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Antons Mislevics", email = "" }
]
requires-python = ">=3.10"
dependencies = [
    "artifacts-keyring>=0.4.0",
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.9.0",
    "jmespath>=1.0.1",
    "keyring>=25.5.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.10.2",
    "pyyaml>=6.0.2",
    "tabulate>=0.9.0",
    "typer>=0.14.0",
]

[project.scripts]
se = "safe_env.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/safe_env/version.py"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B"]

# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# 4. Ignore `E402` (import violations) in all `__init__.py` files, and in selected subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
# 5. Use single quotes in `ruff format`.
quote-style = "single"

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]